+++ SonOfRobin.Core/Game/Pieces/AlchemyLab.cs
@@ -297,7 +297,7 @@ namespace SonOfRobin
                 object buffVal = buff.value;
                 Type valueType = buffVal.GetType();
 
-                float valueMultiplier = 1f + (float)(this.world.random.NextDouble() * ((brewLevel - 1) / 2));
+                float valueMultiplier = 1f + (float)(this.world.random.NextSingle() * ((brewLevel - 1) / 2));
 
                 if (valueType == typeof(byte)) buffVal = (byte)Math.Round((byte)buffVal * valueMultiplier, 0);
                 else if (valueType == typeof(short)) buffVal = (short)Math.Round((short)buffVal * valueMultiplier, 0);
@@ -306,7 +306,7 @@ namespace SonOfRobin
                 else if (valueType == typeof(double)) buffVal = (double)Math.Round((double)buffVal * valueMultiplier, 1);
                 else throw new ArgumentException($"Unsupported valueType for buff type {buff.type} - {valueType}.");
 
-                float durationMultiplier = 1f + (float)(this.world.random.NextDouble() * ((brewLevel - 1) / 2));
+                float durationMultiplier = 1f + (float)(this.world.random.NextSingle() * ((brewLevel - 1) / 2));
 
                 Buff adjustedBuff = new Buff(type: buff.type, value: buffVal, autoRemoveDelay: (int)(buff.autoRemoveDelay * durationMultiplier), isPermanent: buff.isPermanent, canKill: buff.canKill, increaseIDAtEveryUse: buff.increaseIDAtEveryUse);
